We've made some progress on implementing collision with the ground when moving in the DIR_DOWN direction, however we still have a problem where once the player stopes it is possile to press s and have the player move down past the border characters that the code just stopped the player at. This of course will need to be resolved.

+++ game/src/include/common.hpp
@@ -10,8 +10,8 @@ struct yx
     this->y = y;
     this->x = x;
   }
-  size_t y;
-  size_t x;
+  int y;
+  int x;
 };
